% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{antiPrefix}
\alias{antiPrefix}
\title{Select varibles not matching prefix}
\usage{
antiPrefix(.data, prefix)
}
\arguments{
\item{prefix}{}
}
\description{
Select varibles not matching prefix
}
\examples{
antiPrefix(iris, Sepal)
}
