% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{\%!match\%}
\alias{\%!match\%}
\title{Select variables not containing pattern}
\usage{
.data \%!match\% pattern
}
\arguments{
\item{pattern}{}
}
\description{
Select variables not containing pattern
}
\examples{
iris \%!match\% Len
}
