% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{\%!pre\%}
\alias{\%!pre\%}
\title{Select varibles not matching prefix}
\usage{
.data \%!pre\% prefix
}
\arguments{
\item{prefix}{}
}
\description{
Select varibles not matching prefix
}
\examples{
iris \%!pre\% Sepal
}
