% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{\%!suf\%}
\alias{\%!suf\%}
\title{Select variables not ending with pattern}
\usage{
.data \%!suf\% suffix
}
\arguments{
\item{suffix}{}
}
\description{
Select variables not ending with pattern
}
\examples{
iris \%!suf\% Length
}
