% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{\%pby\%}
\alias{\%pby\%}
\title{split dataset by prefix of variable
pby is short for "prefix by"}
\usage{
.data \%pby\% prefix
}
\arguments{
\item{prefix}{the funciton assumes the prefix uniquely identifies the variable}
}
\description{
split dataset by prefix of variable
pby is short for "prefix by"
}
\examples{
iris \%#by\% Spe
iris \%#by\% Spe \%#by\% Petal.Len

}
