% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{\%pre\%}
\alias{\%pre\%}
\title{Select Variable by prefixoperator}
\usage{
.data \%pre\% prefix
}
\arguments{
\item{.data}{array-like object, or list}

\item{prefix}{prefix to find variables by}
}
\description{
Select Variable by prefixoperator
}
\examples{
iris \%pre\% Sepal
}
