% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{\%rgx\%}
\alias{\%rgx\%}
\title{Select variables matching regular expression}
\usage{
.data \%rgx\% regex
}
\arguments{
\item{regex}{}
}
\description{
Select variables matching regular expression
}
\examples{
iris \%rgx\% "(Sepal|Petal).(Length)"

}
