% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{prefix}
\alias{prefix}
\title{Select Variable by prefix}
\usage{
prefix(.data, prefix)
}
\arguments{
\item{.data}{array-like object, or list}

\item{prefix}{prefix to find variables by}
}
\description{
Select Variable by prefix
}
\examples{
cars\%pre\%spe
}
