% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOs.R
\name{prefix}
\alias{prefix}
\title{Select variables matching prefix}
\usage{
prefix(.data, prefix)
}
\arguments{
\item{.data}{array-like object, or list}

\item{prefix}{prefix to find variables by}
}
\description{
Select variables matching prefix
}
\examples{
prefix(iris, Sepal)
}
